chore: create and config `.bundle/config`
1. run `bundle config set path 'vendor/bundle'`;
2. run `bundle config set force_ruby_platform true`;
3. modify `.gitignore` file to contain:
    ```
    /.bundle/*
    !/.bundle/config
    ```
4. run `bundle update --bundler && bundle update && bundle clean --force`.

Note: When both an RVM gemset and `.bundle/config` with `BUNDLE_PATH: "vendor/bundle"` are configured, Bundler prioritizes the `BUNDLE_PATH` setting in `.bundle/config`. As a result, gems will be installed in the `vendor/bundle` directory within the project, overriding the default behavior of installing gems in the RVM gemset directory.